+++ sheepyu-module-system/sheepyu-module-system-biz/src/main/java/top/sheepyu/module/system/service/log/SystemAccessLogService.java
@@ -20,5 +20,5 @@ public interface SystemAccessLogService extends IServiceX<SystemAccessLog> {
 
     PageResult<SystemAccessLog> pageAccessLog(@Valid SystemAccessLogQueryVo queryVo);
 
-    List<Integer> countByWeek(Date beginWeek, Date endWeek);
+    List<Long> countByWeek(Date beginWeek, Date endWeek);
 }
