+++ sheepyu-module-system/sheepyu-module-system-biz/src/main/java/top/sheepyu/module/system/service/log/SystemAccessLogServiceImpl.java
@@ -54,7 +54,7 @@ public class SystemAccessLogServiceImpl extends ServiceImplX<SystemAccessLogMapp
     }
 
     @Override
-    public List<Integer> countByWeek(Date beginWeek, Date endWeek) {
+    public List<Long> countByWeek(Date beginWeek, Date endWeek) {
         return baseMapper.countByWeek(beginWeek, endWeek);
     }
 }
